% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{add_cell_number}
\alias{add_cell_number}
\title{Title}
\usage{
add_cell_number(cci_dt, cond1, cond2, metadata)
}
\arguments{
\item{cci_dt}{x}

\item{cond1}{x}

\item{cond2}{x}

\item{metadata}{x}
}
\value{
x
}
\description{
Title
}
