% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{aggregate_cells_slow}
\alias{aggregate_cells_slow}
\title{Aggregate data per cell-types and conditions}
\usage{
aggregate_cells_slow(expr_tr, metadata, cell_types, cond)
}
\arguments{
\item{expr_tr}{(transposed) expression matrix}

\item{metadata}{x}

\item{cell_types}{x}

\item{cond}{x}
}
\value{
A matrix with the mean expression of each gene for each cell-type
}
\description{
Aggregate data per cell-types and conditions
}
