% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{aggregate_LR}
\alias{aggregate_LR}
\title{Create a dataframe with LR pairs from our 4 sources}
\usage{
aggregate_LR(one2one = TRUE, cpdb = FALSE)
}
\arguments{
\item{one2one}{logical indicating the orthology relationship used during conversion}

\item{cpdb}{logical indicating if returning cpdb}
}
\value{
data.frame with LR pairs and to which sources they belong
}
\description{
Create a dataframe with LR pairs from our 4 sources
}
