% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{aggregate_means}
\alias{aggregate_means}
\title{Compute mean expression per cell-type}
\usage{
aggregate_means(data, group)
}
\arguments{
\item{data}{Data matrix}

\item{group}{Vector indicating groups}
}
\value{
A matrix with the mean expression of each gene for each cell-type
}
\description{
Compute mean expression per cell-type
}
