% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis_2.R
\name{build_cci_array_2}
\alias{build_cci_array_2}
\title{Title}
\usage{
build_cci_array_2(averaged_expr, LR_df)
}
\arguments{
\item{averaged_expr}{x}

\item{LR_df}{x}
}
\value{
x
}
\description{
Title
}
