% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{build_cci_drate_dt}
\alias{build_cci_drate_dt}
\title{Title}
\usage{
build_cci_drate_dt(
  averaged_expr,
  template_cci_dt,
  cond1,
  cond2,
  detection_thr,
  cci_or_drate
)
}
\arguments{
\item{averaged_expr}{x}

\item{template_cci_dt}{x}

\item{cond1}{x}

\item{cond2}{x}

\item{detection_thr}{x}

\item{cci_or_drate}{x}
}
\value{
x
}
\description{
Title
}
