% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{build_celltype_bipartite_graph}
\alias{build_celltype_bipartite_graph}
\title{Title}
\usage{
build_celltype_bipartite_graph(object, disperse = FALSE, dir = NULL)
}
\arguments{
\item{object}{x}

\item{disperse}{x}

\item{dir}{x}
}
\value{
x
}
\description{
Title
}
