% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{build_drate_array}
\alias{build_drate_array}
\title{Title}
\usage{
build_drate_array(detection_rate, detection_thr, LR_df)
}
\arguments{
\item{detection_rate}{x}

\item{detection_thr}{x}

\item{LR_df}{x}
}
\value{
x
}
\description{
Title
}
