% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{build_LR6db}
\alias{build_LR6db}
\title{Build a data.table of curated ligand-receptor interactions obtained from 6 databases.}
\usage{
build_LR6db()
}
\value{
A data.table with ligands, receptors and some annotations (database of origin and source of curation).
}
\description{
Build a data.table of curated ligand-receptor interactions obtained from 6 databases.
}
