% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_LRdb.R
\name{build_LRdb}
\alias{build_LRdb}
\title{Build a data.table of curated ligand-receptor interactions obtained from 6 databases.}
\usage{
build_LRdb(species)
}
\arguments{
\item{species}{human or mouse}
}
\value{
A data.table with ligands, receptors and some annotations (database of origin and source of curation).
}
\description{
Build a data.table of curated ligand-receptor interactions obtained from 6 databases.
}
