% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{BuildNetwork}
\alias{BuildNetwork}
\alias{BuildNetwork.scDiffCom}
\alias{BuildNetwork.scDiffComCombined}
\title{xxx}
\usage{
BuildNetwork(object, ...)

\method{BuildNetwork}{scDiffCom}(
  object,
  network_type = c("condition1_network", "condition2_network", "difference_network",
    "up_regulated_network", "down_regulated_network", "ORA_network"),
  layout_type = c("conventional", "bipartite"),
  abbreviation_table = NULL,
  ...
)

\method{BuildNetwork}{scDiffComCombined}(
  object,
  network_type = c("condition1_network", "condition2_network", "difference_network",
    "up_regulated_network", "down_regulated_network", "ORA_network"),
  layout_type = c("conventional", "bipartite"),
  ID,
  abbreviation_table,
  ...
)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}

\item{network_type}{xxx}

\item{layout_type}{xxx}

\item{abbreviation_table}{xxx}

\item{ID}{xxx}
}
\value{
xxx
}
\description{
xxx
}
