% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{BuildShiny}
\alias{BuildShiny}
\alias{BuildShiny,scDiffCom-method}
\title{A shiny app to display scDiffCom results}
\usage{
BuildShiny(object, reduced_go_table = NULL, ...)

\S4method{BuildShiny}{scDiffCom}(object, reduced_go_table = NULL, ...)
}
\arguments{
\item{object}{\code{scDiffCom} object}

\item{reduced_go_table}{If \code{NULL} (default), over-represented GO terms
are displayed as dot plots. If the output of
 \code{scDiffCom::ReduceGO(object)}, GO terms are displayed on a on treemap
 based on their semantic similarity and over-representation score}

\item{...}{Additional parameters to \code{shiny::runApp}}
}
\value{
Launch a shiny app
}
\description{
Launch a shiny app to explore scDiffCom results
}
