% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_converted_genes_data.R
\docType{data}
\name{CellChat_converted_genes}
\alias{CellChat_converted_genes}
\title{A data-table with genes that need to be converted from CELLCHAT.}
\format{
A 2-column data-table.
}
\usage{
data(CellChat_converted_genes)
}
\description{
A data-table with genes that need to be converted from CELLCHAT.
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
