% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Title}
\usage{
clean_colnames(cci_dt, cond1, cond2, permutation_analysis)
}
\arguments{
\item{cci_dt}{x}

\item{cond1}{x}

\item{cond2}{x}

\item{permutation_analysis}{x}
}
\value{
x
}
\description{
Title
}
