% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Combine_scDiffCom}
\alias{Combine_scDiffCom}
\title{Combine several scDiffCom objects together.}
\usage{
Combine_scDiffCom(l, object_name, verbose)
}
\arguments{
\item{l}{A list of scDiffCom objects.}

\item{object_name}{The name of the combined object that will be created.}

\item{verbose}{Should messages be printed?}
}
\value{
An object of class scDiffComCombined
}
\description{
Combine several scDiffCom objects together.
}
