% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{compute_LR_drate}
\alias{compute_LR_drate}
\title{Compute LR detection rate for one LR-pair over all cell-types}
\usage{
compute_LR_drate(ligand, receptor, detection_rate, detection_thr)
}
\arguments{
\item{ligand}{Name of the ligand}

\item{receptor}{Name of the receptor}

\item{average_expression}{A matrix with the detection rate of a each gene per cell-type}
}
\value{
A matrix with the detection rate of LR for each cell-type pair combination
}
\description{
Compute LR detection rate for one LR-pair over all cell-types
}
