% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{compute_LR_drate_slow}
\alias{compute_LR_drate_slow}
\title{Compute LR detection rate for one LR-pair over all cell-types}
\usage{
compute_LR_drate_slow(ligand, receptor, detection_rate, detection_thr)
}
\arguments{
\item{ligand}{character indicating the name of the ligand}

\item{receptor}{character indicating the name of the receptor}

\item{detection_rate}{x}

\item{detection_thr}{x}
}
\value{
matrix with the detection rate of LR for each cell-type pair combination
}
\description{
Compute LR detection rate for one LR-pair over all cell-types
}
