% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis_2.R
\name{compute_LR_score_2}
\alias{compute_LR_score_2}
\title{Compute LR scores for one LR-pair over all cell-types}
\usage{
compute_LR_score_2(ligand, receptor, averaged_expr)
}
\arguments{
\item{ligand}{character indicating the name of the ligand}

\item{receptor}{character indicating the name of the receptor}

\item{averaged_expr}{matrix with the mean expression of a each gene per cell-type}
}
\value{
matrix with the score of LR for each cell-type pair combination
}
\description{
Compute LR scores for one LR-pair over all cell-types
}
