% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cellphonedb.R
\name{create_cpdp_cci}
\alias{create_cpdp_cci}
\title{Title}
\usage{
create_cpdp_cci(input_dir, condition_id = NULL, cond1 = NULL, cond2 = NULL)
}
\arguments{
\item{input_dir}{x}

\item{condition_id}{x}

\item{cond1}{x}

\item{cond2}{x}
}
\value{
x
}
\description{
Title
}
