% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR6db.R
\name{create_LR_cpdb}
\alias{create_LR_cpdb}
\title{#' Create a data.table with the ligand-receptor interactions from NICHENET.
#'
#' @param one2one logical indicating if the orthology conversion has to be limited to one2one homology; default is FALSE.
#'
#' @return data.table with ligand-receptor interactions and their relevant properties obtained from NICHENET.
prepare_LR_nichenet <- function(
  one2one = FALSE
) {
  LR <- nichenetr::lr_network
  data.table::setDT(LR)
  data.table::setnames(
    x = LR,
    old = c("from", "to", "source"),
    new = c("L1", "R1", "SOURCE")
  )
  ortho <- get_orthologs(
    genes = unique(c(LR$L1, LR$R1)),
    input_species = "human",
    one2one = one2one
  )
  ortho <- stats::na.omit(ortho)
  LR <- merge_LR_orthologs(
    LR_dt = LR,
    ortho_dt = ortho,
    nL = 1,
    charL = "L",
    nR = 1,
    charR = "R"
  )
  cols_to_keep <- c(
    "LR_SORTED", "SOURCE",
    "LIGAND_1", "RECEPTOR_1",
    "LIGAND_1_CONF", "RECEPTOR_1_CONF",
    "LIGAND_1_TYPE", "RECEPTOR_1_TYPE"
  )
  return(LR[, cols_to_keep, with = FALSE])
}
Convert CellphoneDB database in a dataframe either complex or deconvoluted.}
\usage{
create_LR_cpdb(deconvoluted = FALSE)
}
\arguments{
\item{deconvoluted}{logical indicating if returning the complex LR with 1 to 3 genes per category or the 1:1 deconvoluted data.frame.}
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from CELLPHONEDB.
}
\description{
#' Create a data.table with the ligand-receptor interactions from NICHENET.
#'
#' @param one2one logical indicating if the orthology conversion has to be limited to one2one homology; default is FALSE.
#'
#' @return data.table with ligand-receptor interactions and their relevant properties obtained from NICHENET.
prepare_LR_nichenet <- function(
  one2one = FALSE
) {
  LR <- nichenetr::lr_network
  data.table::setDT(LR)
  data.table::setnames(
    x = LR,
    old = c("from", "to", "source"),
    new = c("L1", "R1", "SOURCE")
  )
  ortho <- get_orthologs(
    genes = unique(c(LR$L1, LR$R1)),
    input_species = "human",
    one2one = one2one
  )
  ortho <- stats::na.omit(ortho)
  LR <- merge_LR_orthologs(
    LR_dt = LR,
    ortho_dt = ortho,
    nL = 1,
    charL = "L",
    nR = 1,
    charR = "R"
  )
  cols_to_keep <- c(
    "LR_SORTED", "SOURCE",
    "LIGAND_1", "RECEPTOR_1",
    "LIGAND_1_CONF", "RECEPTOR_1_CONF",
    "LIGAND_1_TYPE", "RECEPTOR_1_TYPE"
  )
  return(LR[, cols_to_keep, with = FALSE])
}
Convert CellphoneDB database in a dataframe either complex or deconvoluted.
}
