% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{distributions<-}
\alias{distributions<-}
\title{Modify scDiffCom \code{distributions}}
\usage{
distributions(x) <- value
}
\arguments{
\item{x}{An scDiffCom object.}

\item{value}{A new list of distributions.}
}
\description{
Modify scDiffCom \code{distributions}
}
