% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{distributions}
\alias{distributions}
\alias{distributions,scDiffCom-method}
\title{Title}
\usage{
distributions(x)

\S4method{distributions}{scDiffCom}(x)
}
\arguments{
\item{x}{x}
}
\value{
x
}
\description{
Title
}
