% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_cell_types}
\alias{filter_cell_types}
\title{Return cell-types containing a minimum of cells in each condition}
\usage{
filter_cell_types(metadata, min_cells)
}
\arguments{
\item{metadata}{Result from prepare_seurat_metadata}

\item{min_cells}{Numeric indicating minimal number of cells in cell-types}
}
\value{
A vector of cell-types with at least min_cells cells
}
\description{
Return cell-types containing a minimum of cells in each condition
}
