% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis_2.R
\name{filter_detected_LR_2}
\alias{filter_detected_LR_2}
\title{Indicates which LR has been detected in at least one CCI}
\usage{
filter_detected_LR_2(array_dr1, array_dr2, LR_df)
}
\arguments{
\item{array_dr1}{3D array for noCond or cond1}

\item{array_dr2}{3D array for cond2 (NULL if noCond)}

\item{LR_df}{LR data.frame}
}
\value{
logical vector
}
\description{
Indicates which LR has been detected in at least one CCI
}
