% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{FilterCCI}
\alias{FilterCCI}
\alias{FilterCCI,scDiffCom-method}
\alias{FilterCCI,scDiffComCombined-method}
\title{xxx}
\usage{
FilterCCI(
  object,
  new_threshold_quantile_score = NULL,
  new_threshold_p_value_specificity = NULL,
  new_threshold_p_value_de = NULL,
  new_threshold_logfc = NULL,
  skip_ora = FALSE,
  verbose = TRUE
)

\S4method{FilterCCI}{scDiffCom}(
  object,
  new_threshold_quantile_score = NULL,
  new_threshold_p_value_specificity = NULL,
  new_threshold_p_value_de = NULL,
  new_threshold_logfc = NULL,
  skip_ora = FALSE,
  verbose = TRUE
)

\S4method{FilterCCI}{scDiffComCombined}(
  object,
  new_threshold_quantile_score = NULL,
  new_threshold_p_value_specificity = NULL,
  new_threshold_p_value_de = NULL,
  new_threshold_logfc = NULL,
  skip_ora = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{xxx}

\item{new_threshold_quantile_score}{xxx}

\item{new_threshold_p_value_specificity}{xxx}

\item{new_threshold_p_value_de}{xxx}

\item{new_threshold_logfc}{xxx}

\item{skip_ora}{xxx}

\item{verbose}{Should messages be printed?}
}
\value{
xxx
}
\description{
xxx
}
