% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_ontology_level}
\alias{gene_ontology_level}
\title{All gene ontology terms annotated with their levels}
\format{
A data.table
}
\usage{
data(gene_ontology_level)
}
\description{
This data.table contains all GO terms retrieved from the package
ontoProc. Each term is annotated with its number of ancestors,
parents and children, as well as with its level (i.e. depth) in
the gene ontology graph. Levels are computed by scDiffCom according to
\code{scDiffCom:::get_GO_LEVELS()}.
}
\references{
\href{https://doi.org/doi:10.18129/B9.bioc.ontoProc}{ontoProc}
}
\keyword{datasets}
