% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{generate_interactive_network}
\alias{generate_interactive_network}
\title{#' Generate interactive networks with visNetwork
#'
#' @param object scDiffCom
#' @param network_type chr in "bipartite", "cells"
#'
#' @return visNetwork interactive network
#' @export}
\usage{
generate_interactive_network(object, network_type)
}
\description{
#' Generate interactive networks with visNetwork
#'
#' @param object scDiffCom
#' @param network_type chr in "bipartite", "cells"
#'
#' @return visNetwork interactive network
#' @export
}
