% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_cci_detected}
\alias{get_cci_detected}
\title{Return scDiffCom \code{cci_detected}}
\usage{
get_cci_detected(x)
}
\arguments{
\item{x}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{cci_detected}
}
