% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{get_cci_score}
\alias{get_cci_score}
\title{Title}
\usage{
get_cci_score(data_t, metadata, LR_data, use_thr, detection_thr, condition_id)
}
\arguments{
\item{data_t}{x}

\item{metadata}{x}

\item{LR_data}{x}

\item{use_thr}{x}

\item{detection_thr}{x}

\item{condition_id}{x}
}
\value{
x
}
\description{
Title
}
