% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_cci_table_detected}
\alias{get_cci_table_detected}
\title{Return scDiffCom \code{cci_table_detected}}
\usage{
get_cci_table_detected(object)
}
\arguments{
\item{object}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{cci_table_detected}
}
