% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_cci_table_filtered}
\alias{get_cci_table_filtered}
\alias{get_cci_table_filtered,scDiffCom-method}
\title{Title}
\usage{
get_cci_table_filtered(x)

\S4method{get_cci_table_filtered}{scDiffCom}(x)
}
\arguments{
\item{x}{x}
}
\value{
x
}
\description{
Title
}
