% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_cci_table_raw}
\alias{get_cci_table_raw}
\title{Return scDiffCom \code{cci_table_raw}}
\usage{
get_cci_table_raw(object)
}
\arguments{
\item{object}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{cci_table_raw}
}
