% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{get_diff_score_fast}
\alias{get_diff_score_fast}
\title{Title}
\usage{
get_diff_score_fast(data_t, metadata, LR_data, permutation_test = TRUE)
}
\arguments{
\item{data_t}{x}

\item{metadata}{x}

\item{LR_data}{x}

\item{permutation_test}{x}
}
\value{
x
}
\description{
Title
}
