% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_distr}
\alias{get_distr}
\alias{distributions<-,scDiffCom-method}
\title{Title}
\usage{
distributions(x) <- value

\S4method{distributions}{scDiffCom}(x) <- value
}
\arguments{
\item{x}{x}

\item{value}{x}
}
\value{

}
\description{
Title
}
