% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_human_orthologs}
\alias{get_human_orthologs}
\title{Convert mouse gene Symbols to human ortholog Symbols.}
\usage{
get_human_orthologs(genes)
}
\arguments{
\item{genes}{A character vector of mouse gene Symbols.}
}
\value{
A dataframe with mouse and human Symbols as columns.
}
\description{
Only keeps genes with orthologogy_confidence == 1
and "ortholog_one2one" mapping. The function deals
with none 1:1 mapping as (to explain). Is it to
stringent? E.g. CDKN2A is not returned...
}
