% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_ora_default}
\alias{get_ora_default}
\title{Return scDiffCom \code{ora_default}}
\usage{
get_ora_default(x)
}
\arguments{
\item{x}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{ora_default}
}
