% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_ora_stringent}
\alias{get_ora_stringent}
\title{Return scDiffCom \code{ora_stringent}}
\usage{
get_ora_stringent(x)
}
\arguments{
\item{x}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{ora_stringent}
}
