% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_ora_table}
\alias{get_ora_table}
\title{Return scDiffCom \code{ora_table}}
\usage{
get_ora_table(object)
}
\arguments{
\item{object}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{ora_table}
}
