% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_ora_tables}
\alias{get_ora_tables}
\alias{get_ora_tables,scDiffCom-method}
\title{Title}
\usage{
get_ora_tables(x)

\S4method{get_ora_tables}{scDiffCom}(x)
}
\arguments{
\item{x}{x}
}
\value{
x
}
\description{
Title
}
