% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_param}
\alias{get_param}
\alias{parameters<-,scDiffCom-method}
\title{Title}
\usage{
parameters(x) <- value

\S4method{parameters}{scDiffCom}(x) <- value
}
\arguments{
\item{x}{x}

\item{value}{x}
}
\value{
x
}
\description{
Title
}
