% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetDistributions}
\alias{GetDistributions}
\alias{GetDistributions,scDiffCom-method}
\title{Return the slot \code{distributions} from a scDiffCom object}
\usage{
GetDistributions(object)

\S4method{GetDistributions}{scDiffCom}(object)
}
\arguments{
\item{object}{\code{scDiffCom} object}
}
\description{
Return the slot \code{distributions} from a scDiffCom object
}
