% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetParameters}
\alias{GetParameters}
\alias{GetParameters,scDiffComBase-method}
\title{Return the slot \code{parameters} from a scDiffCom object}
\usage{
GetParameters(object)

\S4method{GetParameters}{scDiffComBase}(object)
}
\arguments{
\item{object}{\code{scDiffCom} object}
}
\value{
A list of parameters.
}
\description{
Return the parameters that have been passed to
 \code{\link{run_interaction_analysis}} as well as a few other
 parameters computed alongside the analysis.
}
