% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetTableCCI}
\alias{GetTableCCI}
\alias{GetTableCCI,scDiffCom-method}
\title{Return information from one of the scDiffCom CCI data.tables}
\usage{
GetTableCCI(object, type, simplified)

\S4method{GetTableCCI}{scDiffCom}(object, type = c("detected", "raw"), simplified = TRUE)
}
\arguments{
\item{object}{\code{scDiffCom} object}

\item{type}{Table to extract information from. Can be either
\code{"detected"} (default) or \code{"raw"}.}

\item{simplified}{If \code{TRUE} (default) only the most
informative columns of the data.table are returned.}
}
\value{
A data.table
}
\description{
Return information from one of the scDiffCom CCI data.tables
}
