% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetTableORA}
\alias{GetTableORA}
\alias{GetTableORA,scDiffCom-method}
\title{Return information from some of the scDiffCom ORA data.tables}
\usage{
GetTableORA(object, categories, simplified)

\S4method{GetTableORA}{scDiffCom}(object, categories = "all", simplified = TRUE)
}
\arguments{
\item{object}{\code{scDiffCom} object}

\item{categories}{Names of the categories to return. If \code{"all"}
(default), returns all of them.}

\item{simplified}{If \code{TRUE} (default) only the most
informative columns of the data.table are returned.}
}
\value{
A list of data.tables
}
\description{
Return information from some of the scDiffCom ORA data.tables
}
