% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{is_detected}
\alias{is_detected}
\title{#' Compute arithmetic mean
#'
#' @param x
#' @param y
#' @param scale_factor
#'
#' @return
#' @export
#'
#' @examples
cci_amean <- Vectorize(function(
  x,
  y,
  scale_factor = 0
) {
  m <- (x+y)/2
  if(!(scale_factor == 0)) {
    return(m/(m+scale_factor))
  } else {
    return(m)
  }
})
Title}
\usage{
is_detected(x, y, thr)
}
\arguments{
\item{x}{x}

\item{y}{x}

\item{thr}{x}
}
\value{
xxx
}
\description{
#' Compute arithmetic mean
#'
#' @param x
#' @param y
#' @param scale_factor
#'
#' @return
#' @export
#'
#' @examples
cci_amean <- Vectorize(function(
  x,
  y,
  scale_factor = 0
) {
  m <- (x+y)/2
  if(!(scale_factor == 0)) {
    return(m/(m+scale_factor))
  } else {
    return(m)
  }
})
Title
}
