% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{is_detected_full}
\alias{is_detected_full}
\title{Determine if the CCI interaction is detected or not}
\usage{
is_detected_full(x_dr, x_ncells, y_dr, y_ncells, dr_thr, min_cells)
}
\arguments{
\item{x_dr}{x}

\item{x_ncells}{x}

\item{y_dr}{x}

\item{y_ncells}{x}

\item{dr_thr}{x}

\item{min_cells}{x}
}
\value{
x
}
\description{
Determine if the CCI interaction is detected or not
}
