% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR6db_data.R
\docType{data}
\name{LR6db}
\alias{LR6db}
\title{Ligand-receptor data.table with subunits interactions based on 6 literature sources.}
\format{
A data.table with columns including the (subunit) ligand/receptor genes and some relevant properties.
}
\usage{
data(LR6db)
}
\description{
LR6db contains mouse LR interactions converted from human genes when necessary.
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
