% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR6db_wSource_data.R
\docType{data}
\name{LR6db_wSource}
\alias{LR6db_wSource}
\title{Ligand-receptor interactions from 6 databases.}
\format{
A list of 1 data.table and one list.
}
\usage{
data(LR6db_wSource)
}
\description{
LR6db_wSource contains a data.table of LR interactions and a list of the sources from each database.
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
