% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRall_data.R
\docType{data}
\name{LRall}
\alias{LRall}
\title{Ligand-Receptor data.frames based on 4 literature sources.}
\format{
A list of 2 data.frames.
}
\usage{
data(LRall)
}
\description{
LRall_one2one (resp. LRall_many2many) contains LR pairs converted from mouse to human genes based on one2one (resp. many2many)
orthology relationship.
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
