% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRdb_human}
\alias{LRdb_human}
\title{A database of human ligand-receptor interactions.}
\format{
A list of 3 data.tables: a non-curated table, a curated table and a table of corresponding GO terms
}
\usage{
data(LRdb_human)
}
\description{
The DB has been compiled from eight external studies.
}
\details{
Can be obtained by running scDiffCom:::build_LRdb(species = "human")
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
