% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{merge_LR_orthologs}
\alias{merge_LR_orthologs}
\title{Take a LR data.table of human genes and convert it with mouse genes.}
\usage{
merge_LR_orthologs(LR_dt, ortho_dt, nL, charL, nR, charR)
}
\arguments{
\item{LR_dt}{data.table of human ligand-recetpor interactions.}

\item{ortho_dt}{data.table of human-mouse homologues.}

\item{nL}{numeric indicating the number of ligand columns in LR_dt.}

\item{charL}{character indicating the prefix name of the ligand columns in LR_dt.}

\item{nR}{numeric indicating the number of receptor columns in LR_dt.}

\item{charR}{character indicating the prefix name of the receptor columns in LR_dt.}
}
\value{
data.table with the mouse ligand-receptor interactions. Filtering of duplicated is also performed.
}
\description{
Take a LR data.table of human genes and convert it with mouse genes.
}
