% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{parameters<-}
\alias{parameters<-}
\title{Modify scDiffCom \code{parameters}}
\usage{
parameters(x) <- value
}
\arguments{
\item{x}{An scDiffCom object.}

\item{value}{A new list of parameters.}
}
\description{
Modify scDiffCom \code{parameters}
}
