% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{parameters}
\alias{parameters}
\alias{parameters,scDiffCom-method}
\title{Return the slot 'parameters'}
\usage{
parameters(x)

\S4method{parameters}{scDiffCom}(x)
}
\arguments{
\item{x}{scDiffCom object}
}
\value{

}
\description{
Return the slot 'parameters'
}
