% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{plot_ora}
\alias{plot_ora}
\title{Title}
\usage{
plot_ora(
  object,
  category,
  OR_val,
  pval_val,
  ORA_score_val,
  max_value,
  OR_cutoff = 1,
  pval_cutoff = 0.05
)
}
\arguments{
\item{object}{x}

\item{category}{x}

\item{OR_val}{x}

\item{pval_val}{x}

\item{ORA_score_val}{x}

\item{max_value}{x}

\item{OR_cutoff}{x}

\item{pval_cutoff}{x}
}
\value{
x
}
\description{
Title
}
