% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{PlotORA}
\alias{PlotORA}
\title{#' Perform filtering and over-representation analysis on intercellular communication patterns
#'
#' This function is called internally by \code{run_interaction_analysis} after the permutation tests
#'  have been performed. Based on the threshold parameters, it returns detected and differentially expressed
#'  CCIs in the slot \code{cci_detected} and results of over-representation analysis in \code{cci_ora_default}.
#'  The function can be run with new threshold parameters on any scDiffCom object that already contain the slot
#'  \code{cci_raw}. This allows the user to test various filtering parameters without the need to rerun the
#'  potentially time-consuming permutation analysis. When new thresholds are defined the slot \code{parameters} is
#'  modified accordingly.
#'  Filtering and over-representation are not independent as the second depends on the first. Therefore, when
#'  running filtering with new parameters, we also need to update the ORA results.
#'
#' @param object An scDiffCom object previously returned by \code{run_interaction_analysis}.
#' @param new_threshold_quantile_score A new threshold for the quantile score. Set to \code{NULL} by default.
#' @param new_threshold_p_value_specificity A new threshold for the specificity p-value specificity. Set to \code{NULL} by default.
#' @param new_threshold_p_value_de A new threshold for the differential expression p-value. Set to \code{NULL} by default.
#' @param new_threshold_logfc A new threshold for the differential expression logfc. Set to \code{NULL} by default.
#' @param skip_ora Should the over-representation analysis be skipped? Set to \code{FALSE} by default.
#'  Setting it to \code{FALSE} might be useful if we are interested in performing a rapid parameter scan only on the filtering
#'  parameters. Note that in such case, the slot \code{ora_default} is returned empty.
#' @param verbose Should messages be printed?
#'
#' @return An S4 object of class \code{scDiffCom}.
#' @export
Plot over-represented terms of a given category}
\usage{
PlotORA(
  object,
  category,
  regulation,
  max_terms_show,
  OR_threshold = 1,
  p_value_threshold = 0.05,
  stringent = FALSE
)
}
\arguments{
\item{object}{An scDiffCom object.}

\item{category}{The category over wich the ORA has been performed (e.g. "GO_TERMS").}

\item{regulation}{One from "UP", "DOWN", "DIFF" or "FLAT".}

\item{max_terms_show}{The maximal number of terms to show on the plot.}

\item{OR_threshold}{Only display terms with an odds ratio bigger than this threshold. Set to \code{1} by default.}

\item{p_value_threshold}{Only display terms with a BH-p-value smaller thatn this threshold. Set to \code{0.05} by default.}

\item{stringent}{Should we display the default ORA results or the stringent ORA results. Set to \code{FALSE} by default.}
}
\value{
A ggplot object.
}
\description{
#' Perform filtering and over-representation analysis on intercellular communication patterns
#'
#' This function is called internally by \code{run_interaction_analysis} after the permutation tests
#'  have been performed. Based on the threshold parameters, it returns detected and differentially expressed
#'  CCIs in the slot \code{cci_detected} and results of over-representation analysis in \code{cci_ora_default}.
#'  The function can be run with new threshold parameters on any scDiffCom object that already contain the slot
#'  \code{cci_raw}. This allows the user to test various filtering parameters without the need to rerun the
#'  potentially time-consuming permutation analysis. When new thresholds are defined the slot \code{parameters} is
#'  modified accordingly.
#'  Filtering and over-representation are not independent as the second depends on the first. Therefore, when
#'  running filtering with new parameters, we also need to update the ORA results.
#'
#' @param object An scDiffCom object previously returned by \code{run_interaction_analysis}.
#' @param new_threshold_quantile_score A new threshold for the quantile score. Set to \code{NULL} by default.
#' @param new_threshold_p_value_specificity A new threshold for the specificity p-value specificity. Set to \code{NULL} by default.
#' @param new_threshold_p_value_de A new threshold for the differential expression p-value. Set to \code{NULL} by default.
#' @param new_threshold_logfc A new threshold for the differential expression logfc. Set to \code{NULL} by default.
#' @param skip_ora Should the over-representation analysis be skipped? Set to \code{FALSE} by default.
#'  Setting it to \code{FALSE} might be useful if we are interested in performing a rapid parameter scan only on the filtering
#'  parameters. Note that in such case, the slot \code{ora_default} is returned empty.
#' @param verbose Should messages be printed?
#'
#' @return An S4 object of class \code{scDiffCom}.
#' @export
Plot over-represented terms of a given category
}
