% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR6db.R
\name{prepare_LR_CellChat}
\alias{prepare_LR_CellChat}
\title{Create a data.table with the ligand-receptor interactions from CellChat.}
\usage{
prepare_LR_CellChat()
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from CELLCHAT.
}
\description{
Create a data.table with the ligand-receptor interactions from CellChat.
}
