% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR6db.R
\name{prepare_LR_nichenet}
\alias{prepare_LR_nichenet}
\title{Create a data.table with the ligand-receptor interactions from NICHENET.}
\usage{
prepare_LR_nichenet(one2one = FALSE)
}
\arguments{
\item{one2one}{logical indicating if the orthology conversion has to be limited to one2one homology; default is FALSE.}
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from NICHENET.
}
\description{
Create a data.table with the ligand-receptor interactions from NICHENET.
}
