% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{prepare_LR_scsr}
\alias{prepare_LR_scsr}
\title{Create a data.table with the ligand-receptor interactions from SingleCellSignalR.}
\usage{
prepare_LR_scsr(one2one = FALSE)
}
\arguments{
\item{one2one}{logical indicating if the orthology conversion has to be limited to one2one homology; default is FALSE.}
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from SingleCellSignalR.
}
\description{
Create a data.table with the ligand-receptor interactions from SingleCellSignalR.
}
