% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{prepare_LR_scTensor}
\alias{prepare_LR_scTensor}
\title{Create a data.table with the ligand-receptor interactions from scTensor.}
\usage{
prepare_LR_scTensor()
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from scTensor.
}
\description{
Create a data.table with the ligand-receptor interactions from scTensor.
}
