% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prepare_seurat_data}
\alias{prepare_seurat_data}
\title{Prepare Seurat data matrix for downstream analysis.}
\usage{
prepare_seurat_data(seurat_obj, assay = "RNA", slot = "data",
  log_scale = TRUE, convert_to_human = FALSE, return_type = "dense")
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{assay}{character indicating the assay to choose from}

\item{slot}{character indicating the slot to choose from}

\item{log_scale}{logical indicating if returning log-normalized value (only for slot data)}

\item{convert_to_human}{logical indicating if gene names have to be converted to human orthologs}

\item{return_type}{character indicating the class of the return data (sparse, dense or data.frame)}
}
\value{
a dgCMatrix, a matrix or a data.frame
}
\description{
Prepare Seurat data matrix for downstream analysis.
}
\examples{
prepare_seurat_data(seurat_random_test)
}
