% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prepare_seurat_metadata}
\alias{prepare_seurat_metadata}
\title{Prepare Seurat metadata for downstream analysis}
\usage{
prepare_seurat_metadata(seurat_obj, seurat_cell_type_id,
  condition_id = NULL)
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{seurat_cell_type_id}{a character indicating the column of cell-types in the Seurat object}

\item{condition_id}{a character indicating a column with some condition on the cells}
}
\value{
a dataframe
}
\description{
Prepare Seurat metadata for downstream analysis
}
\examples{
prepare_seurat_metadata(seurat_random_test)
}
