% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{prepare_template_cci}
\alias{prepare_template_cci}
\title{Create a template data.table with all the CCIs}
\usage{
prepare_template_cci(LR_df, cell_types)
}
\arguments{
\item{LR_df}{data.frame of LR CCI}

\item{cell_types}{character of the cell_types present in the Seurat object}
}
\value{
a data.table
}
\description{
Create a template data.table with all the CCIs
}
