% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preprocess_LR}
\alias{preprocess_LR}
\title{Correspondance between dataset genes and LR-pairs}
\usage{
preprocess_LR(data, LR_data)
}
\arguments{
\item{data}{matrix or data.frame with genes as rows and cells as columns.}

\item{LR_data}{3-column data.frame of Ligand-Receptor pairs,
satisfying colnames(LR_data) == c("GENESYMB_L", "GENESYMB_R", "SYMB_LR").}
}
\value{
A list containing the subsetted data and the subsetted LR-pairs.
}
\description{
Correspondance between dataset genes and LR-pairs
}
