% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preprocess_seurat}
\alias{preprocess_seurat}
\title{Return data and metadata of a Seurat object for downstream analysis.}
\usage{
preprocess_seurat(
  seurat_object,
  assay,
  slot,
  log_scale,
  convert_to_human,
  return_type,
  seurat_cell_type_id,
  condition_id,
  min_cells
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{assay}{character indicating the assay of the Seurat object where to pull the data from; default is "RNA".}

\item{slot}{character indicating the slot of the Seurat object where to pull the data from; default is "data".}

\item{log_scale}{logical indicating if using log-normalized data (TRUE) or normalized data (FALSE); default is "TRUE".
Only considered if slot == "data".}

\item{convert_to_human}{logical indicating if converting mouse gene names to human gene names.}

\item{return_type}{character indicating the class of the returned data (sparse, dense or data.frame)}

\item{seurat_cell_type_id}{character indicating the column specifying the cell-types of the cells}

\item{condition_id}{character indicating the column specifying the conditions on the cells. Set to NULL to run the analysis
on no condition; default is NULL.}

\item{min_cells}{numeric indicating the minimal number of cells each cluster need to contain to be considered in the analysis.}
}
\value{
A list containing data, metadata, cell-types and possibly orthology mapping.
}
\description{
Return data and metadata of a Seurat object for downstream analysis.
}
