% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_cci_analysis}
\alias{run_cci_analysis}
\title{Main internal routine}
\usage{
run_cci_analysis(
  expr_tr,
  metadata,
  template_cci_dt,
  threshold,
  condition_id,
  differential_analysis,
  specificity_analysis,
  iterations,
  one_sided,
  return_distr = FALSE
)
}
\arguments{
\item{expr_tr}{matrix with the (transposed) expression data from the Seurat object}

\item{metadata}{data.table with the relevant metadata from the Seurat object}

\item{template_cci_dt}{data.table with a template of all CCIs}

\item{threshold}{numeric indicating the percentage of cells that need to express a gene in a cluster for the gene to be
considered detected.}

\item{condition_id}{character indicating the column specifying the conditions on the cells. Set to NULL to run the analysis
on no condition; default is NULL.}

\item{differential_analysis}{logical indicating if performing the permutation test for the significance of differential
expression between the conditions. Only considered when condidition_id is not NULL; default is TRUE.}

\item{specificity_analysis}{logical indicating if performing the permutation test for the specificity of each CCI on each condition;
default is FALSE.}

\item{iterations}{integer indicating the number of iterations during the permutation test.}

\item{one_sided}{logical indicating if computing differential p-values from a one-sided or two-sided test.}

\item{return_distr}{logical indicating if returning the matrix with the distributions obtained from the permutation test.}
}
\value{
x
}
\description{
Main internal routine
}
