% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_wrapper.R
\name{run_cpdb_dot_plot}
\alias{run_cpdb_dot_plot}
\title{Title}
\usage{
run_cpdb_dot_plot(
  means_path = NULL,
  pvalues_path = NULL,
  output_path = NULL,
  output_name = NULL,
  rows = NULL,
  columns = NULL,
  verbose = TRUE
)
}
\arguments{
\item{means_path}{x}

\item{pvalues_path}{x}

\item{output_path}{x}

\item{output_name}{x}

\item{rows}{x}

\item{columns}{x}

\item{verbose}{x}
}
\value{
x
}
\description{
Title
}
