% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cellphonedb.R
\name{run_cpdb_from_files}
\alias{run_cpdb_from_files}
\title{Run CellPhoneDB from text files.}
\usage{
run_cpdb_from_files(
  data_path,
  metadata_path,
  method = "statistical_analysis",
  project_name = NULL,
  iterations = NULL,
  threshold = NULL,
  result_precision = NULL,
  counts_data = NULL,
  output_path = NULL,
  output_format = NULL,
  means_result_name = NULL,
  significant_means_result_name = NULL,
  deconvoluted_result_name = NULL,
  verbose = TRUE,
  pvalues_result_name = NULL,
  debug_seed = NULL,
  threads = NULL,
  subsampling = FALSE,
  subsampling_log = FALSE,
  subsampling_num_pc = NULL,
  subsampling_num_cells = NULL
)
}
\arguments{
\item{data_path}{a text file}

\item{metadata_path}{a text file}

\item{method}{character}

\item{project_name}{character}

\item{iterations}{integer}

\item{threshold}{numeric}

\item{result_precision}{numeric}

\item{counts_data}{character, type of gene identifier}

\item{output_path}{character}

\item{output_format}{character}

\item{means_result_name}{character}

\item{significant_means_result_name}{character}

\item{deconvoluted_result_name}{character}

\item{verbose}{logical}

\item{pvalues_result_name}{character}

\item{debug_seed}{numeric}

\item{threads}{numeric}

\item{subsampling}{logical}

\item{subsampling_log}{logical}

\item{subsampling_num_pc}{numeric}

\item{subsampling_num_cells}{numeric}
}
\value{
Run CellPhoneDB and return results in output directory
}
\description{
Run CellPhoneDB from text files.
}
\examples{

}
