% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_wrapper.R
\name{run_cpdb_from_seurat}
\alias{run_cpdb_from_seurat}
\title{Run CellPhoneDB from a Seurat object}
\usage{
run_cpdb_from_seurat(
  seurat_object,
  assay = "RNA",
  slot = "data",
  log_scale = FALSE,
  celltype_column_id,
  input_species = "mouse",
  min_cells = 5,
  condition_column_id = NULL,
  input_dir = getwd(),
  create_plots = FALSE,
  method = "statistical_analysis",
  iterations = NULL,
  threshold = NULL,
  result_precision = NULL,
  counts_data = NULL,
  output_format = NULL,
  verbose = TRUE,
  debug_seed = NULL,
  threads = NULL,
  subsampling = FALSE,
  subsampling_log = FALSE,
  subsampling_num_pc = NULL,
  subsampling_num_cells = NULL,
  return_full_dt = TRUE
)
}
\arguments{
\item{seurat_object}{x}

\item{assay}{x}

\item{slot}{x}

\item{log_scale}{x}

\item{celltype_column_id}{x}

\item{input_species}{x}

\item{min_cells}{x}

\item{condition_column_id}{x}

\item{input_dir}{x}

\item{create_plots}{x}

\item{method}{x}

\item{iterations}{x}

\item{threshold}{x}

\item{result_precision}{x}

\item{counts_data}{x}

\item{output_format}{x}

\item{verbose}{x}

\item{debug_seed}{x}

\item{threads}{x}

\item{subsampling}{x}

\item{subsampling_log}{x}

\item{subsampling_num_pc}{x}

\item{subsampling_num_cells}{x}

\item{return_full_dt}{x}
}
\value{
Return the results of CellPhoneDB
}
\description{
Run CellPhoneDB from a Seurat object
}
