% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cellphonedb.R
\name{run_cpdb_heatmap}
\alias{run_cpdb_heatmap}
\title{Title}
\usage{
run_cpdb_heatmap(
  metadata_path,
  pvalues_path = NULL,
  output_path = NULL,
  count_name = NULL,
  log_name = NULL,
  count_network_name = NULL,
  interaction_count_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{metadata_path}{x}

\item{pvalues_path}{x}

\item{output_path}{x}

\item{count_name}{x}

\item{log_name}{x}

\item{count_network_name}{x}

\item{interaction_count_name}{x}

\item{verbose}{x}
}
\value{
x
}
\description{
Title
}
