% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_diffcom}
\alias{run_diffcom}
\title{Determine (differential) intercellular communication scores (with significance)}
\usage{
run_diffcom(
  seurat_object,
  LR_data,
  seurat_cell_type_id = "cell_ontology_class",
  condition_id = NULL,
  assay = "RNA",
  slot = "data",
  log_scale = TRUE,
  min_cells = 5,
  threshold = 0.1,
  permutation_analysis = TRUE,
  one_sided = FALSE,
  iterations = 1000,
  return_distr = FALSE,
  seed = 42
)
}
\arguments{
\item{seurat_object}{Seurat object of several cell-types and human/mouse genes.}

\item{LR_data}{3-column data.frame of Ligand-Receptor pairs,
satisfying colnames(LR_data) == c("GENESYMB_L", "GENESYMB_R", "SYMB_LR").}

\item{seurat_cell_type_id}{character indicating the column specifying the cell-types of the cells}

\item{condition_id}{character indicating the column specifying the conditions on the cells. Set to NULL to run the analysis
on no condition; default is NULL.}

\item{assay}{character indicating the assay of the Seurat object where to pull the data from; default is "RNA".}

\item{slot}{character indicating the slot of the Seurat object where to pull the data from; default is "data".}

\item{log_scale}{logical indicating if using log-normalized data (TRUE) or normalized data (FALSE); default is "TRUE".
Only considered if slot == "data".}

\item{min_cells}{numeric indicating the minimal number of cells each cluster need to contain to be considered in the analysis.}

\item{threshold}{numeric indicating the percentage of cells that need to express a gene in a cluster for the gene to be
considered detected.}

\item{permutation_analysis}{logical indicating if performing permutation test; default is TRUE.}

\item{one_sided}{logical indicating if computing differential p-values from a one-sided or two-sided test.}

\item{iterations}{integer indicating the number of iterations during the permutation test.}

\item{return_distr}{logical indicating if returning the matrix with the distributions obtained from the permutation test.}

\item{seed}{numeric indicating a seeding value for reproducibility.}
}
\value{
A data.table where each row is CCI. The columns vary in functions of the parameters used when calling the function.
It includes the CCI information and for each condition the scores, detection rates and possibly p-values.
}
\description{
Read a Seurat object and a data.frame of Ligand-Receptor interactions.
Compute the score of each cell-cell interaction (CCI) for one or two conditions on the cells.
There is the possibility to return a specicifity p-value for each CCI over the conditions.
When two conditions are considered, it computes the changes of each CCI with the option to return a p-value to assess
the significance of each change.
Statistical assessements are based on a permutation test (either permuting the condition or cluster labels.)
}
