% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_diffcom_from_seurat}
\alias{run_diffcom_from_seurat}
\title{Title}
\usage{
run_diffcom_from_seurat(
  seurat_obj,
  LR_data,
  assay = "RNA",
  slot = "data",
  log_scale = TRUE,
  seurat_cell_type_id = "cell_ontology_class",
  condition_id = NULL,
  min_cells = 5,
  threshold = 0.1,
  convert_to_human = FALSE,
  statistical_analysis = TRUE,
  iterations = 1000
)
}
\arguments{
\item{seurat_obj}{x}

\item{LR_data}{x}

\item{assay}{x}

\item{slot}{x}

\item{log_scale}{x}

\item{seurat_cell_type_id}{x}

\item{condition_id}{x}

\item{min_cells}{x}

\item{threshold}{x}

\item{convert_to_human}{x}

\item{statistical_analysis}{x}

\item{iterations}{x}
}
\value{
x
}
\description{
Title
}
