% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_analysis.R
\name{run_interaction_analysis}
\alias{run_interaction_analysis}
\title{Run (differential) intercellular communication analysis}
\usage{
run_interaction_analysis(
  seurat_object,
  LRdb_species,
  seurat_celltype_id,
  seurat_condition_id,
  iterations = 1000,
  scdiffcom_object_name = "scDiffCom_object",
  seurat_assay = "RNA",
  seurat_slot = "data",
  log_scale = FALSE,
  score_type = "geometric_mean",
  threshold_min_cells = 5,
  threshold_pct = 0.1,
  threshold_quantile_score = 0.2,
  threshold_p_value_specificity = 0.05,
  threshold_p_value_de = 0.05,
  threshold_logfc = log(1.5),
  return_distributions = FALSE,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{Object of class \emph{Seurat} that contains normalized data and
relevant \code{meta.data} annotations. Gene names must be MGI (mouse) or HGNC (human).}

\item{LRdb_species}{Either "mouse" or "human". Specifies which internal ligand-receptor database to use.}

\item{seurat_celltype_id}{Column name of \emph{Seurat} \code{meta.data} that indicates the cell-type
of each cell (e.g.: \code{"CELL_TYPE"}).}

\item{seurat_condition_id}{Length-3 named list that indicates 1) \code{"column_name"} of
\emph{Seurat} \code{meta.data} that indicates to which of two groups each cell belongs, 2) \code{cond1_name}
and 3) \code{cond2_name} (e.g.: \code{list(column_name = "AGE", cond1_name = "YOUNG", cond2_name = "OLD")}).
By convention, LOGFC are positive for \code{score(cond2) > score(cond1)}. Set to \code{NULL} to only
perform a detection analysis withoug looking for differences between conditions.}

\item{iterations}{Number of permutations to perform during the statistical analysis. Default is \code{1000}.
Can be set to \code{0} to quickly return partial results without performing the statistical analysis.}

\item{scdiffcom_object_name}{Name of the returned \emph{scDiffCom} object. Default is "scDiffCom_object". Having
different names for different objects is useful to combine them in an \emph{scDiffComCombined} object.}

\item{seurat_assay}{\emph{Seurat} assay from which to extract the data. Default is "RNA".}

\item{seurat_slot}{\emph{Seurat} slot from which to extract the data. Default is "data".}

\item{log_scale}{Whether to use log-normalized data or non-log-normalized data. Default is \code{FALSE}.}

\item{score_type}{Either "geometric_mean" (default) or "arithmetic_mean". THe metric to compute the CCI score.}

\item{threshold_min_cells}{Used to filter out the CCIs for which either the ligand(s) or receptor(s)
are expressed in less cells than this threshold. Cell-types that have less cells that the threshold are removed
initially. Default is \code{5}.}

\item{threshold_pct}{Used to filter out the CCIs for which either the ligand(s) or receptor(s)
are expressed in less than this threshold fraction of the cells. Default is \code{0.1}.}

\item{threshold_quantile_score}{Quantile value used to filter interactions with low scores. Default is
\code{0.2}. Can be modified at will after statistical analysis.}

\item{threshold_p_value_specificity}{Maximal BH p-value to consider an interaction to be specific. Default
is \code{0.05}. Can be modified at will after statistical analysis.}

\item{threshold_p_value_de}{Maximal BH p-value to consider an interaction to be differentially expressed.
Default is \code{0.05}. Can be modified at will after statistical analysis.}

\item{threshold_logfc}{Minimal LOGFC (natural log scale) to consider an interaction to be differentially
expressed. Default is \code{log(1.5)}. Can be modified at will after statistical analysis.}

\item{return_distributions}{Wheter to return the null distributions from the permutation test. Only available
when computing less than or 1000 \code{iterations}. Default is \code{FALSE}.}

\item{seed}{Set a random seed. Default is \code{42}.}

\item{verbose}{Whether to print messages. Default is \code{TRUE}.}
}
\value{
An S4 object of class \code{scDiffCom}.
}
\description{
Perform (differential) cell-cell interaction analysis based on scRNA-seq data (Seurat object)
 and an internal database of ligand-receptor interactions.
}
