% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{run_ora}
\alias{run_ora}
\title{Run over-representation analysis on the results from scDiffCom}
\usage{
run_ora(
  object,
  verbose = TRUE,
  logfc_threshold = NULL,
  categories = c("LR_CELLTYPE", "LR_NAME", "GO"),
  overwrite = FALSE
)
}
\arguments{
\item{object}{The list of results returned by run_scdiffcom}

\item{verbose}{Print messages}

\item{logfc_threshold}{Log fold-change threshold (in absolute value) to be used to define the categories of interest.
Default to cutoff_logfc used for the filtering analysis.}

\item{categories}{The categories over which the test is done}

\item{overwrite}{Should existing categories be overwriten in case they match with new categories?}
}
\value{
A data.table
}
\description{
Run over-representation analysis on the results from scDiffCom
}
