% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_permutation}
\alias{run_permutation}
\title{Title}
\usage{
run_permutation(data_t, metadata, LR_data, iterations)
}
\arguments{
\item{data_t}{x}

\item{metadata}{x}

\item{LR_data}{x}

\item{iterations}{x}
}
\value{
x
}
\description{
Title
}
