% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_simple_analysis}
\alias{run_simple_analysis}
\title{Compute the score and detection rate of each CCI}
\usage{
run_simple_analysis(
  expr_tr,
  metadata,
  template_cci_dt,
  cond1,
  cond2,
  threshold,
  min_cells = min_cells,
  compute_fast
)
}
\arguments{
\item{expr_tr}{matrix with the (transposed) expression data from the Seurat object}

\item{metadata}{data.table with the relevant metadata from the Seurat object}

\item{template_cci_dt}{data.table with a template of all CCIs}

\item{cond1}{character indicating the name of the first condition, or NULL if no condition}

\item{cond2}{character indicating the name of the first condition, or NULL if no condition}

\item{threshold}{numeric indicating the percentage of cells that need to express a gene in a cluster for the gene to be
considered detected.}

\item{min_cells}{x}

\item{compute_fast}{logical indicating if doing a fast computation for the permutation test or a full computation with
detection rate.}
}
\value{
a data.table or a (list of) numeric vector(s)
}
\description{
Compute the score and detection rate of each CCI
}
