% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_stat_analysis}
\alias{run_stat_analysis}
\title{Title}
\usage{
run_stat_analysis(
  expr_tr,
  metadata,
  cell_types,
  LR_df,
  cond1,
  cond2,
  iterations,
  use_case
)
}
\arguments{
\item{expr_tr}{x}

\item{metadata}{x}

\item{cell_types}{x}

\item{LR_df}{x}

\item{cond1}{x}

\item{cond2}{x}

\item{iterations}{x}

\item{use_case}{x}
}
\value{
x
}
\description{
Title
}
