% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_stat_iteration}
\alias{run_stat_iteration}
\title{Title}
\usage{
run_stat_iteration(expr_tr, metadata, template_cci_dt, cond1, cond2, use_case)
}
\arguments{
\item{expr_tr}{x}

\item{metadata}{x}

\item{template_cci_dt}{x}

\item{cond1}{x}

\item{cond2}{x}

\item{use_case}{x}
}
\value{
x
}
\description{
Title
}
