% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RunORA}
\alias{RunORA}
\alias{RunORA.scDiffCom}
\alias{RunORA.scDiffComCombined}
\title{Perform over-representation analysis on detected intercellular communication patterns}
\usage{
RunORA(object, ...)

\method{RunORA}{scDiffCom}(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS", "KEGG_PWS"),
  overwrite = TRUE,
  verbose = TRUE,
  ...
)

\method{RunORA}{scDiffComCombined}(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS"),
  overwrite = TRUE,
  global = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An scDiffCom object previously returned by \code{run_interaction_analysis}.}

\item{...}{Arguments passed to other methods}

\item{categories}{A character vector specifying the categories on which to perform the analysis. One data.table is returned
for each category. Set to \code{c("ER_CELLTYPES", "LRI", "GO_TERMS")} by default.}

\item{overwrite}{Should existing categories be overwritten in case they match with new categories?}

\item{verbose}{Should messages be printed?}

\item{global}{xxx}
}
\value{
An S4 object of class \code{scDiffCom} with updated slots \code{ora_defaut} (and potentially \code{ora_stringent}).
}
\description{
Over-representation analysis performed when an scDiffCom object contains results
 obtained from differential expression analysis.
}
