% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{scDiffCom-class}
\alias{scDiffCom-class}
\alias{scDiffCom}
\alias{parameters,scDiffCom-method}
\alias{parameters<-,scDiffCom-method}
\alias{get_cci_raw,scDiffCom-method}
\alias{set_cci_raw,scDiffCom-method}
\alias{get_cci_detected,scDiffCom-method}
\alias{set_cci_detected,scDiffCom-method}
\alias{distributions,scDiffCom-method}
\alias{distributions<-,scDiffCom-method}
\alias{get_ora_default,scDiffCom-method}
\alias{set_ora_default,scDiffCom-method}
\alias{get_ora_stringent,scDiffCom-method}
\alias{set_ora_stringent,scDiffCom-method}
\title{The scDiffCom Class}
\usage{
\S4method{parameters}{scDiffCom}(x)

\S4method{parameters}{scDiffCom}(x) <- value

\S4method{get_cci_raw}{scDiffCom}(x)

\S4method{set_cci_raw}{scDiffCom}(x, new_cci_raw)

\S4method{get_cci_detected}{scDiffCom}(x)

\S4method{set_cci_detected}{scDiffCom}(x, new_cci_detected)

\S4method{distributions}{scDiffCom}(x)

\S4method{distributions}{scDiffCom}(x) <- value

\S4method{get_ora_default}{scDiffCom}(x)

\S4method{set_ora_default}{scDiffCom}(x, new_ora_default)

\S4method{get_ora_stringent}{scDiffCom}(x)

\S4method{set_ora_stringent}{scDiffCom}(x, new_ora_stringent)
}
\arguments{
\item{x}{An scDiffCom object.}

\item{value}{A new list of parameters.}

\item{new_cci_raw}{A \code{cci_raw} data.table.}

\item{new_cci_detected}{A \code{cci_detected} data.table.}

\item{new_ora_default}{A \code{ora_default} list of data.tables.}

\item{new_ora_stringent}{A \code{ora_stringent} list of data.tables.}
}
\description{
The scDiffCom object stores the results of the intercellular communication analysis
 performed by \code{run_interaction_analysis} on a single-cell Seurat object.
}
\section{Functions}{
\itemize{
\item \code{parameters,scDiffCom-method}: Return scDiffCom \code{parameters}.

\item \code{parameters<-,scDiffCom-method}: Modify scDiffCom \code{parameters}.

\item \code{get_cci_raw,scDiffCom-method}: Return scDiffCom \code{cci_raw}.

\item \code{set_cci_raw,scDiffCom-method}: Modify scDiffCom \code{cci_raw}.

\item \code{get_cci_detected,scDiffCom-method}: Return scDiffCom \code{cci_detected}.

\item \code{set_cci_detected,scDiffCom-method}: Modify scDiffCom \code{cci_detected}.

\item \code{distributions,scDiffCom-method}: Return scDiffCom \code{distributions}.

\item \code{distributions<-,scDiffCom-method}: Modify scDiffCom \code{distributions}.

\item \code{get_ora_default,scDiffCom-method}: Return scDiffCom \code{ora_default}.

\item \code{set_ora_default,scDiffCom-method}: Modify scDiffCom \code{ora_default}.

\item \code{get_ora_stringent,scDiffCom-method}: Return scDiffCom \code{ora_stringent}.

\item \code{set_ora_stringent,scDiffCom-method}: Modify scDiffCom \code{ora_stringent}.
}}

\section{Slots}{

\describe{
\item{\code{parameters}}{A list of the parameters passed to \code{run_interaction_analysis}.}

\item{\code{cci_raw}}{A data.table of all possible cell-cell interactions (CCIs); namely with \eqn{n^2 \cdot m} rows,
where n is the number of cell-types and m the number of ligand-receptor pairs.}

\item{\code{cci_detected}}{A data.table of the detected CCIs obtained from the raw data.table after filtering.}

\item{\code{distributions}}{A list of matrices that contain the distributions over each CCI obtained from the permutation test(s).
Storing such distributions will make the object very heavy for more than 1000 permutations. It is only intended to be used for
benchmarking or debugging.}

\item{\code{ora_default}}{A list of data.tables (one per category) storing the results of the over-representation analysis
performed on the \code{cci_detected} table.}

\item{\code{ora_stringent}}{Same as \code{ora_default} but with a more stringent logfc threshold to specifically focus on strong signals.}
}}

