% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{scDiffComCombined-class}
\alias{scDiffComCombined-class}
\title{The scDiffComCombined Class}
\description{
A scDiffComCombined object stores the results of multiple scDiffCom
  objects at a single place.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{A list of the parameters used to build the object.}

\item{\code{cci_table_raw}}{A data.table with all possible CCIs induced from
the original Seurat object and the internal LRI database.}

\item{\code{cci_table_detected}}{A data.table with only the detected CCIs obtained
after filtering.}

\item{\code{ora_table}}{A data.table with the results of the
over-representation analysis.}
}}

