% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{set_cci_detected}
\alias{set_cci_detected}
\title{Modify scDiffCom \code{cci_detected}}
\usage{
set_cci_detected(x, new_cci_detected)
}
\arguments{
\item{x}{An scDiffCom object.}

\item{new_cci_detected}{A \code{cci_detected} data.table.}
}
\description{
Modify scDiffCom \code{cci_detected}
}
