% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{set_cci_raw}
\alias{set_cci_raw}
\title{Modify scDiffCom \code{cci_raw}}
\usage{
set_cci_raw(x, new_cci_raw)
}
\arguments{
\item{x}{An scDiffCom object.}

\item{new_cci_raw}{A \code{cci_raw} data.table.}
}
\description{
Modify scDiffCom \code{cci_raw}
}
