% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_cci_table_filtered}
\alias{set_cci_table_filtered}
\alias{set_cci_table_filtered,scDiffCom-method}
\title{Title}
\usage{
set_cci_table_filtered(x, new_cci_table_filtered)

\S4method{set_cci_table_filtered}{scDiffCom}(x, new_cci_table_filtered)
}
\arguments{
\item{x}{x}

\item{new_cci_table_filtered}{x}
}
\value{
x
}
\description{
Title
}
