% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_cci_table_raw}
\alias{set_cci_table_raw}
\alias{set_cci_table_raw,scDiffCom-method}
\title{Title}
\usage{
set_cci_table_raw(x, new_cci_table_raw)

\S4method{set_cci_table_raw}{scDiffCom}(x, new_cci_table_raw)
}
\arguments{
\item{x}{x}

\item{new_cci_table_raw}{x}
}
\value{
x
}
\description{
Title
}
