% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{set_ora_default}
\alias{set_ora_default}
\title{Modify scDiffCom \code{ora_default}}
\usage{
set_ora_default(x, new_ora_default)
}
\arguments{
\item{x}{An scDiffCom object.}

\item{new_ora_default}{A \code{ora_default} list of data.tables.}
}
\description{
Modify scDiffCom \code{ora_default}
}
