% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{set_ora_stringent}
\alias{set_ora_stringent}
\title{Modify scDiffCom \code{ora_stringent}}
\usage{
set_ora_stringent(x, new_ora_stringent)
}
\arguments{
\item{x}{An scDiffCom object.}

\item{new_ora_stringent}{A \code{ora_stringent} list of data.tables.}
}
\description{
Modify scDiffCom \code{ora_stringent}
}
