% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_ora_tables}
\alias{set_ora_tables}
\alias{set_ora_tables,scDiffCom-method}
\title{Title}
\usage{
set_ora_tables(x, new_ora_tables)

\S4method{set_ora_tables}{scDiffCom}(x, new_ora_tables)
}
\arguments{
\item{x}{x}

\item{new_ora_tables}{x}
}
\value{
x
}
\description{
Title
}
