% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_random_test_data.R
\docType{data}
\name{seurat_random_test}
\alias{seurat_random_test}
\title{A random Seurat file for testing}
\format{
An object of class \code{"Seurat"}.
}
\usage{
data(seurat_random_test)
}
\description{
A random Seurat file for testing
}
\keyword{datasets}
