% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_sample_tms_liver}
\alias{seurat_sample_tms_liver}
\title{A (sampled) Seurat object from the Tabula Muris Senis liver.}
\format{
A Seurat object.
}
\usage{
data(seurat_sample_tms_liver)
}
\description{
This is only a sample from the original object. It only contains
two cell-types and roughly 2000 genes. The slot "data" contains
log-normalized values with normalization performed before the down-sampling.
It is only intended to be used as a toy model to show the functionalities of the package.
It does not convey meaningful biological information.
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
