% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{show,scDiffCom-method}
\alias{show,scDiffCom-method}
\title{Display a scDiffCom object}
\usage{
\S4method{show}{scDiffCom}(object)
}
\arguments{
\item{object}{\code{scDiffCom} object}
}
\description{
Display a scDiffCom object
}
