% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_preparation.R
\name{subset_by_LR}
\alias{subset_by_LR}
\title{Select LR genes present in the dataset}
\usage{
subset_by_LR(data, LR_data)
}
\arguments{
\item{data}{A matrix or data.frame with genes as rows and cells as columns}

\item{LR_data}{A data.frame with one LR pair per row}
}
\value{
A list containing the subsetted data and the subsetted LR-pairs
}
\description{
Select LR genes present in the dataset
}
