% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{aggregate_cells}
\alias{aggregate_cells}
\title{Aggregate data per cell-types and conditions}
\usage{
aggregate_cells(expr_tr, metadata, cell_types, cond)
}
\arguments{
\item{data}{Data matrix}

\item{group}{Vector indicating groups}
}
\value{
A matrix with the mean expression of each gene for each cell-type
}
\description{
Aggregate data per cell-types and conditions
}
