% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{aggregate_cells}
\alias{aggregate_cells}
\title{Title}
\usage{
aggregate_cells(expr_tr, metadata, is_cond)
}
\arguments{
\item{expr_tr}{x}

\item{metadata}{x}

\item{is_cond}{x}
}
\value{
x
}
\description{
Title
}
