% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{BuildNetwork}
\alias{BuildNetwork}
\alias{BuildNetwork.scDiffCom}
\alias{BuildNetwork.scDiffComCombined}
\title{xxx}
\usage{
BuildNetwork(object, ...)

\method{BuildNetwork}{scDiffCom}(object, network_type, ...)

\method{BuildNetwork}{scDiffComCombined}(object, network_type, subobject_name, ...)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
xxx
}
