% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{BuildNetwork}
\alias{BuildNetwork}
\alias{BuildNetwork.scDiffCom}
\alias{BuildNetwork.scDiffComCombined}
\title{xxx}
\usage{
BuildNetwork(object, ...)

\method{BuildNetwork}{scDiffCom}(
  object,
  network_type = c("ORA", "COUNTS_DIFF", "COUNTS_COND1", "COUNTS_COND2"),
  network_layout = c("bipartite", "celltypes"),
  ...
)

\method{BuildNetwork}{scDiffComCombined}(
  object,
  network_type = c("ORA", "COUNTS_DIFF", "COUNTS_COND1", "COUNTS_COND2"),
  network_layout = c("bipartite", "celltypes"),
  ID,
  ...
)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}

\item{network_type}{xxx}

\item{network_layout}{xxx}

\item{ID}{xxx}
}
\value{
xxx
}
\description{
xxx
}
