% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{BuildNetwork}
\alias{BuildNetwork}
\alias{BuildNetwork,scDiffCom-method}
\alias{BuildNetwork,scDiffComCombined-method}
\title{xxx}
\usage{
BuildNetwork(
  object,
  network_type = c("condition1_network", "condition2_network", "difference_network",
    "up_regulated_network", "down_regulated_network", "ORA_network"),
  layout_type = c("conventional", "bipartite"),
  abbreviation_table = NULL,
  ...
)

\S4method{BuildNetwork}{scDiffCom}(
  object,
  network_type = c("condition1_network", "condition2_network", "difference_network",
    "up_regulated_network", "down_regulated_network", "ORA_network"),
  layout_type = c("conventional", "bipartite"),
  abbreviation_table = NULL
)

\S4method{BuildNetwork}{scDiffComCombined}(
  object,
  network_type = c("condition1_network", "condition2_network", "difference_network",
    "up_regulated_network", "down_regulated_network", "ORA_network"),
  layout_type = c("conventional", "bipartite"),
  abbreviation_table,
  ID
)
}
\arguments{
\item{object}{xxx}

\item{network_type}{xxx}

\item{layout_type}{xxx}

\item{abbreviation_table}{xxx}

\item{...}{Extra named arguments passed to BuildNetwork}

\item{ID}{xxx}
}
\value{
xxx
}
\description{
xxx
}
