% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{BuildShiny}
\alias{BuildShiny}
\alias{BuildShiny,scDiffCom-method}
\title{A shiny app to display scDiffCom results}
\usage{
BuildShiny(object, reduce_go = TRUE, ...)

\S4method{BuildShiny}{scDiffCom}(object, reduce_go = TRUE, ...)
}
\arguments{
\item{object}{\code{scDiffCom} object}

\item{reduce_go}{If \code{TRUE} (default), reduce go terms by semantic
similarity and allows shiny to plot GO Treemaps.}

\item{...}{Additional parameters to \code{shiny::runApp}}
}
\value{
Launch a shiny app
}
\description{
Launch a shiny app to explore scDiffCom results
}
