% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{combine_LR_db}
\alias{combine_LR_db}
\title{Create a data.table with the ligand-receptor interactions from 6 databases.}
\usage{
combine_LR_db(one2one = FALSE)
}
\arguments{
\item{one2one}{logical indicating if the orthology conversion has to be limited to one2one homology; default is FALSE.}
}
\value{
data.table with ligand-receptor interactions and relevant properties from 6 databases.
}
\description{
Create a data.table with the ligand-receptor interactions from 6 databases.
}
