% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{compute_LR_score}
\alias{compute_LR_score}
\title{Compute LR scores for one LR-pair over all cell-types}
\usage{
compute_LR_score(ligand, receptor, averaged_expr)
}
\arguments{
\item{ligand}{Name of the ligand}

\item{receptor}{Name of the receptor}

\item{average_expression}{A matrix with the mean expression of a each gene per cell-type}
}
\value{
A matrix with the score of LR for each cell-type pair combination
}
\description{
Compute LR scores for one LR-pair over all cell-types
}
