% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cellphonedb.R
\name{create_cpdb_input}
\alias{create_cpdb_input}
\title{Create text files to be used by CellPhoneDB from a Seurat object}
\usage{
create_cpdb_input(
  seurat_obj,
  assay = "RNA",
  slot = "data",
  log_scale = TRUE,
  seurat_cell_type_id,
  condition_id = NULL,
  input_dir = getwd()
)
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{assay}{assay to pull data from}

\item{slot}{slot to pull data from}

\item{log_scale}{logical}

\item{seurat_cell_type_id}{xx}

\item{condition_id}{xx}

\item{input_dir}{xx}
}
\value{
none
}
\description{
Create text files to be used by CellPhoneDB from a Seurat object
}
\examples{

}
