% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR6db.R
\name{create_LR_cpdb}
\alias{create_LR_cpdb}
\title{Convert CellphoneDB database in a dataframe either complex or deconvoluted.}
\usage{
create_LR_cpdb(deconvoluted = FALSE)
}
\arguments{
\item{deconvoluted}{logical indicating if returning the complex LR with 1 to 3 genes per category or the 1:1 deconvoluted data.frame.}
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from CELLPHONEDB.
}
\description{
Convert CellphoneDB database in a dataframe either complex or deconvoluted.
}
