% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{distributions}
\alias{distributions}
\title{Return scDiffCom \code{distributions}}
\usage{
distributions(object)
}
\arguments{
\item{object}{An scDiffCom object.}
}
\description{
Return scDiffCom \code{distributions}
}
