% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{generate_interactive_network}
\alias{generate_interactive_network}
\title{Generate interactive networks with visNetwork}
\usage{
generate_interactive_network(object, network_type)
}
\arguments{
\item{object}{scDiffCom}

\item{network_type}{chr in "bipartite", "cells"}
}
\value{
visNetwork interactive network
}
\description{
Generate interactive networks with visNetwork
}
