% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_orthologs}
\alias{get_orthologs}
\title{Conversion between mouse and human genes using orthologs.}
\usage{
get_orthologs(genes, input_species)
}
\arguments{
\item{genes}{character vector of the symbols of the genes (mgi_symbol or hgnc_symbol)}

\item{input_species}{character indicating the species of the input genes; either "mouse" or "human".}
}
\value{
2-column data-frame mouse and human orthologs
}
\description{
Conversion is based on orthologogy_confidence == 1
and "ortholog_one2one" mapping.
}
