% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetTableCCI}
\alias{GetTableCCI}
\alias{GetTableCCI,scDiffCom-method}
\alias{GetTableCCI,scDiffComCombined-method}
\title{Return one of the scDiffCom CCI data.tables}
\usage{
GetTableCCI(object, type, simplified)

\S4method{GetTableCCI}{scDiffCom}(object, type = c("detected", "raw"), simplified = TRUE)

\S4method{GetTableCCI}{scDiffComCombined}(object, type = c("detected", "raw"), simplified = TRUE)
}
\arguments{
\item{object}{An object of class `scDiffCom` or `scDiffComCombined`}

\item{type}{Specifies the table to return: either "detected" or "raw".
Default is \code{"detected"}.}

\item{simplified}{Logical indicating if all columns or only the most
informative ones are returned. Default is \code{TRUE}.}
}
\value{
A data.table
}
\description{
Return one of the scDiffCom CCI data.tables
}
