% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{is_detected}
\alias{is_detected}
\title{Determine if two values are bigger than a threshold at the same time.}
\usage{
is_detected(x, y, thr)
}
\arguments{
\item{x}{x}

\item{y}{x}

\item{thr}{x}
}
\value{
x
}
\description{
Determine if two values are bigger than a threshold at the same time.
}
