% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR6db_data.R
\docType{data}
\name{LR6db}
\alias{LR6db}
\title{Ligand-receptor interactions from 6 databases.}
\format{
A list of 2 data.tables and one list.
}
\usage{
data(LR6db)
}
\description{
LR6db contains two data.tables of LR interactions and a list of the sources from each database.
}
\references{
to add
(\href{to_add}{PubMed})
}
\keyword{datasets}
