% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRI_mouse}
\alias{LRI_mouse}
\title{A database of mouse ligand-receptor interactions.}
\format{
A list with seven items:
\enumerate{
  \item LRI_not_curated: a data.table of curated and non-curated LRIs
  \item LRI_curated: a data.table of curated LRIs
  \item LRI_curated_GO: a data.table with GO terms attached to
   curated LRI
  \item LRI_curated_KEGG: a data.table with KEGG pathways attached to
   curated LRIs
   \item LRI_retrieved_dates: dates at which data have been retrieved
   from the eight external databases
  \item LRI_retrieved_from: paths or packages from where data have
   been retrieved
  \item LRI_biomart_ensembl_version: version of ensembl used for
   GO annotation and orthology conversion
}
}
\usage{
data(LRI_mouse)
}
\description{
A list that contains mouse ligand-receptor interactions (LRIs)
 with GO-term and KEGG-pathway annotations. The LRIs have been retrieved
 from eight databases (see References). Curated and non-curated LRIs
 are distinguished (note that only curated LRIs are used in scDiffCom
 interaction analysis, non-curated LRIs are given for reference).
}
\details{
The code to build the database is  scDiffCom:::build_LRI(species = "mouse")
}
\references{
CellChat (\href{https://www.biorxiv.org/content/10.1101/2020.07.21.214387v1}{bioRxiv}),
 CellPhoneDB (\href{https://pubmed.ncbi.nlm.nih.gov/32103204/}{PMID: 32103204}),
 CellTalkDB (\href{https://pubmed.ncbi.nlm.nih.gov/33147626/}{PMID: 33147626}),
 connectomeDB2020 (\href{https://pubmed.ncbi.nlm.nih.gov/33024107/}{PMID: 33024107}),
 ICELLNET (\href{https://pubmed.ncbi.nlm.nih.gov/33597528/}{PMID: 33597528}),
 NicheNet (\href{https://pubmed.ncbi.nlm.nih.gov/31819264/}{PMID: 31819264}),
 SingleCellSignalR (\href{https://pubmed.ncbi.nlm.nih.gov/32196115/}{PMID: 32196115}),
 scTensor (\href{https://www.biorxiv.org/content/10.1101/566182v1}{bioRxiv})
}
\keyword{datasets}
