% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{PlotORA}
\alias{PlotORA}
\alias{PlotORA.scDiffCom}
\alias{PlotORA.scDiffComCombined}
\title{xxx}
\usage{
PlotORA(object, ...)

\method{PlotORA}{scDiffCom}(
  object,
  category,
  regulation,
  max_terms_show,
  OR_threshold = 1,
  p_value_threshold = 0.05,
  ...
)

\method{PlotORA}{scDiffComCombined}(
  object,
  subID,
  category,
  regulation,
  max_terms_show,
  global = FALSE,
  OR_threshold = 1,
  p_value_threshold = 0.05,
  ...
)
}
\arguments{
\item{object}{xxx}

\item{...}{Arguments passed to other methods}

\item{category}{xxx}

\item{regulation}{xxx}

\item{max_terms_show}{xxx}

\item{OR_threshold}{xxx}

\item{p_value_threshold}{xxx}

\item{subID}{xxx}

\item{global}{xxx}
}
\value{
xxx
}
\description{
xxx
}
