% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{PlotORA}
\alias{PlotORA}
\alias{PlotORA,scDiffCom-method}
\alias{PlotORA,scDiffComCombined-method}
\title{xxx}
\usage{
PlotORA(
  object,
  category,
  regulation,
  max_terms_show,
  OR_threshold = 1,
  p_value_threshold = 0.05,
  ...
)

\S4method{PlotORA}{scDiffCom}(
  object,
  category,
  regulation,
  max_terms_show,
  OR_threshold = 1,
  p_value_threshold = 0.05
)

\S4method{PlotORA}{scDiffComCombined}(
  object,
  category,
  regulation,
  max_terms_show,
  OR_threshold = 1,
  p_value_threshold = 0.05,
  subID
)
}
\arguments{
\item{object}{xxx}

\item{category}{xxx}

\item{regulation}{xxx}

\item{max_terms_show}{xxx}

\item{OR_threshold}{xxx}

\item{p_value_threshold}{xxx}

\item{...}{Extra named arguments passed to PlotORA}

\item{subID}{xxx}
}
\value{
xxx
}
\description{
xxx
}
