% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{PlotORA}
\alias{PlotORA}
\title{Plot over-represented terms of a given category}
\usage{
PlotORA(
  object,
  category,
  regulation,
  max_terms_show,
  OR_threshold = 1,
  p_value_threshold = 0.05,
  stringent = FALSE
)
}
\arguments{
\item{object}{An scDiffCom object.}

\item{category}{The category over wich the ORA has been performed (e.g. "GO_TERMS").}

\item{regulation}{One from "UP", "DOWN", "DIFF" or "FLAT".}

\item{max_terms_show}{The maximal number of terms to show on the plot.}

\item{OR_threshold}{Only display terms with an odds ratio bigger than this threshold. Set to \code{1} by default.}

\item{p_value_threshold}{Only display terms with a BH-p-value smaller thatn this threshold. Set to \code{0.05} by default.}

\item{stringent}{Should we display the default ORA results or the stringent ORA results. Set to \code{FALSE} by default.}
}
\value{
A ggplot object.
}
\description{
Plot over-represented terms of a given category
}
