% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{prepare_LR_cpdb}
\alias{prepare_LR_cpdb}
\title{Create a data.table with the ligand-receptor interactions from CELLPHONEDB.}
\usage{
prepare_LR_cpdb(one2one = FALSE, deconvoluted = FALSE)
}
\arguments{
\item{one2one}{logical indicating if the orthology conversion has to be limited to one2one homology; default is FALSE.}

\item{deconvoluted}{logical indiciating if the interactions are deconvoluted when complex; default is FALSE}
}
\value{
data.table with ligand-receptor interactions and their relevant properties obtained from CELLPHONEDB.
}
\description{
Create a data.table with the ligand-receptor interactions from CELLPHONEDB.
}
