% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{prepare_LR_cpdb}
\alias{prepare_LR_cpdb}
\title{Create a LR data.frame based on CellPhoneDB database.}
\usage{
prepare_LR_cpdb(one2one = TRUE, deconvoluted = TRUE)
}
\arguments{
\item{one2one}{logical indicating if using one2one orthology relationship}

\item{deconvoluted}{logical indicating if using 1:1 LR versus complex LR}
}
\value{
data.frame with 3 LR columns and possibly more detailed information
}
\description{
Create a LR data.frame based on CellPhoneDB database.
}
\examples{
\dontrun{
LRcpdb_one2one <- prepare_LR_cpdb(one2one = TRUE, deconvoluted = TRUE)
LRcpdb_many2many <- prepare_LR_cpdb(one2one = FALSE, deconvoluted = TRUE)
}
}
