% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{prepare_LR_nichenet}
\alias{prepare_LR_nichenet}
\title{Create a LR data.frame based on nicheNet database}
\usage{
prepare_LR_nichenet(detailed = FALSE, one2one = TRUE)
}
\arguments{
\item{detailed}{logical indicating if returning only 3 columns or the more detailed data.frame.}

\item{one2one}{logical indicating if using one2one orthology relationship}
}
\value{
data.frame with 3 LR columns and possibly more detailed information
}
\description{
Create a LR data.frame based on nicheNet database
}
\examples{
\dontrun{
LRniche_one2one <- prepare_LR_nichenet(detailed = FALSE, one2one = TRUE)
LRniche_many2many <- prepare_LR_nichenet(detailed = FALSE, one2one = FALSE)
}
}
