% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LR_databases.R
\name{prepare_LR_scTensor}
\alias{prepare_LR_scTensor}
\title{Create a LR data.table based on the scTensor mouse database}
\usage{
prepare_LR_scTensor(detailed = FALSE)
}
\arguments{
\item{detailed}{logical indicating if returning only 3 columns or the more detailed data.frame.}
}
\value{
data.frame with 3 or 6 columns
}
\description{
Create a LR data.table based on the scTensor mouse database
}
\examples{
\dontrun{
LRsct <- prepare_LR_scTensor(detailed = FALSE)
}
}
