% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prepare_seurat_data}
\alias{prepare_seurat_data}
\title{Prepare Seurat data matrix for downstream analysis.}
\usage{
prepare_seurat_data(
  seurat_object,
  assay = "RNA",
  slot = "data",
  log_scale = TRUE,
  convert_to_human = FALSE,
  return_type = "dense"
)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{assay}{character indicating the assay of the Seurat object where to pull the data from; default is "RNA".}

\item{slot}{character indicating the slot of the Seurat object where to pull the data from; default is "data".}

\item{log_scale}{logical indicating if using log-normalized data (TRUE) or normalized data (FALSE); default is "TRUE".
Only considered if slot == "data".}

\item{convert_to_human}{logical indicating if converting mouse gene names to human gene names.}

\item{return_type}{character indicating the class of the return data (sparse, dense or data.frame)}
}
\value{
A list with data as first argument (a dgCMatrix, a matrix or a data.frame) and the gene mapping if converstion to orthologs
}
\description{
Prepare Seurat data matrix for downstream analysis.
}
