% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis_3.R
\name{prepare_template_cci}
\alias{prepare_template_cci}
\title{Title}
\usage{
prepare_template_cci(LR_df, cell_types)
}
\arguments{
\item{LR_df}{x}

\item{cell_types}{x}
}
\value{
x
}
\description{
Title
}
