% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preprocess_LR}
\alias{preprocess_LR}
\title{Correspondance between dataset genes and LR-pairs}
\usage{
preprocess_LR(data, LR_data)
}
\arguments{
\item{data}{x}

\item{LR_data}{x}
}
\value{
A list containing the subsetted data and the subsetted LR-pairs.
}
\description{
Correspondance between dataset genes and LR-pairs
}
