% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_cci_analysis}
\alias{run_cci_analysis}
\title{Title}
\usage{
run_cci_analysis(
  expr_tr,
  metadata,
  cell_types,
  LR_df,
  threshold,
  condition_id,
  statistical_analysis,
  specificity_analysis,
  iterations
)
}
\arguments{
\item{expr_tr}{x}

\item{metadata}{x}

\item{cell_types}{x}

\item{LR_df}{x}

\item{threshold}{x}

\item{condition_id}{x}

\item{statistical_analysis}{x}

\item{specificity_analysis}{x}

\item{iterations}{x}
}
\value{
x
}
\description{
Title
}
