% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cellphonedb.R
\name{run_cpdb_from_seurat}
\alias{run_cpdb_from_seurat}
\title{Run CellPhoneDB from a Seurat object}
\usage{
run_cpdb_from_seurat(seurat_obj, assay = "RNA", slot = "data",
  log_scale = TRUE, seurat_cell_type_id, condition_id = NULL,
  input_dir = getwd(), method = "statistical_analysis",
  project_name = NULL, iterations = NULL, threshold = NULL,
  result_precision = NULL, counts_data = NULL, output_path = NULL,
  output_format = NULL, means_result_name = NULL,
  significant_mean_result_name = NULL, deconvoluted_result_name = NULL,
  verbose = TRUE, pvalues_result_name = NULL, debug_seed = NULL,
  threads = NULL, subsampling = FALSE, subsampling_log = FALSE,
  subsampling_num_pc = NULL, subsampling_num_cells = NULL)
}
\arguments{
\item{seurat_obj}{x}

\item{assay}{x}

\item{slot}{x}

\item{log_scale}{x}

\item{seurat_cell_type_id}{x}

\item{condition_id}{x}

\item{input_dir}{x}

\item{method}{x}

\item{project_name}{x}

\item{iterations}{x}

\item{threshold}{x}

\item{result_precision}{x}

\item{counts_data}{x}

\item{output_path}{x}

\item{output_format}{x}

\item{means_result_name}{x}

\item{significant_mean_result_name}{x}

\item{deconvoluted_result_name}{x}

\item{verbose}{x}

\item{pvalues_result_name}{x}

\item{debug_seed}{x}

\item{threads}{x}

\item{subsampling}{x}

\item{subsampling_log}{x}

\item{subsampling_num_pc}{x}

\item{subsampling_num_cells}{x}
}
\value{
Return the results of CellPhoneDB
}
\description{
Run CellPhoneDB from a Seurat object
}
\examples{

}
