% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_diffcom_from_seurat}
\alias{run_diffcom_from_seurat}
\title{Run the full differential analysis from a Seurat object.}
\usage{
run_diffcom_from_seurat(
  seurat_obj,
  LR_data,
  assay = "RNA",
  slot = "data",
  log_scale = FALSE,
  seurat_cell_type_id = "cell_ontology_class",
  condition_id = NULL,
  min_cells = 5,
  threshold = 0.1,
  convert_to_human = FALSE,
  statistical_analysis = TRUE,
  iterations = 1000,
  return_type = "data.table"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{LR_data}{A ligand-receptor data.frame (a specific format is expected)}

\item{assay}{The Seurat assay to pull data from; default is "RNA"}

\item{slot}{The Seurat slot to pull data from; default is "data"}

\item{log_scale}{Whether to return log-normalized or normalized data (only relevant when slot = "data"); default is TRUE}

\item{seurat_cell_type_id}{Name of the column of the metadata data.frame containing the cell-type ids}

\item{condition_id}{Name of the column of the metadata data.frame containing the the condition on the cells. Set to NULL for no conditions}

\item{min_cells}{Minimum number of cells (per condition if relevant) required to keep a cell-type}

\item{threshold}{Minimum percentage of cells that have to express a gene to consider this gene in the analysis}

\item{convert_to_human}{Whether to convert LR genes from mouse to human; default is FALSE}

\item{statistical_analysis}{Whether to perform the permutation analysis}

\item{iterations}{Number of permutations when runnning the statistical analysis}

\item{return_type}{Wheter to return an array or a data.table}
}
\value{
Depending on return_type, either return an array or a data.table
}
\description{
Run the full differential analysis from a Seurat object.
}
