% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{run_filtering_and_ORA}
\alias{run_filtering_and_ORA}
\title{Run filtering analysis on the result of scDiffCom}
\usage{
run_filtering_and_ORA(
  object,
  verbose = TRUE,
  new_cutoff_quantile_score = NULL,
  new_cutoff_pval_specificity = NULL,
  new_cutoff_pval_de = NULL,
  new_cutoff_logfc = NULL
)
}
\arguments{
\item{object}{The list of results returned by run_scdiffcom}

\item{verbose}{Print messages}

\item{new_cutoff_quantile_score}{New quantile used to define the threshold of detection regarding the score of an interaction.}

\item{new_cutoff_pval_specificity}{New P-value threshold that indicates when an interaction is specific}

\item{new_cutoff_pval_de}{New P-value threshold that indicates when an interaction is differentially expressed}

\item{new_cutoff_logfc}{New LOGFC threshold to filter interactions}
}
\value{
Return a list of result in the scDiffCom format with new filtering analysis.
}
\description{
Run filtering analysis on the result of scDiffCom
}
