% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{run_interaction_analysis}
\alias{run_interaction_analysis}
\title{Run scDiffCom analysis}
\usage{
run_interaction_analysis(
  seurat_object,
  LR_object,
  celltype_column_id,
  condition_column_id,
  cond1_name,
  cond2_name,
  object_name = "scDiffCom_object",
  assay = "RNA",
  slot = "data",
  log_scale = FALSE,
  min_cells = 5,
  pct_threshold = 0.1,
  permutation_analysis = TRUE,
  iterations = 1000,
  cutoff_quantile_score = 0.25,
  cutoff_pval_specificity = 0.05,
  cutoff_pval_de = 0.05,
  cutoff_logfc = log(1.1),
  return_distr = FALSE,
  seed = 42,
  verbose = TRUE,
  sparse = TRUE
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{LR_object}{A data.table with ligand-receptor interactions}

\item{celltype_column_id}{The column of the Seurat meta.data that contains the cell-type of each cell}

\item{condition_column_id}{The column of the Seurat meta.data that contains the group of each cell (only 2 groups allowed).
Use NULL to run the analysis without conditions.}

\item{cond1_name}{Name of the first condition. Use NULL if no condition,
otherwise it should corresponds to one of the groups contained in celltype_column_id.
During differential expression, LOGFC is positive when score(cond1) > score(cond1).}

\item{cond2_name}{Name of the second condition. Similar as cond1_name.}

\item{object_name}{Name of the scDiffCom object that will be returned.}

\item{assay}{Seurat assay from which the data are extracted.}

\item{slot}{Seurat slot from which the data are extracted.}

\item{log_scale}{Logical indicating if working in log-space or not.}

\item{min_cells}{Minimal number of cells for a cell-type to be considered in the analysis.}

\item{pct_threshold}{Minimal percentage of cells required to express a gene, for the gene to be considered in the analysis.}

\item{permutation_analysis}{Logical indicating if performing the statistical analysis (specificity and differential expression).}

\item{iterations}{Number of permutations performed when permutation_analysis == TRUE}

\item{cutoff_quantile_score}{Quantile used to define the threshold of detection regarding the score of an interaction.}

\item{cutoff_pval_specificity}{P-value threshold that indicates when an interaction is specific}

\item{cutoff_pval_de}{P-value threshold that indicates when an interaction is differentially expressed}

\item{cutoff_logfc}{LOGFC threshold to filter interactions}

\item{return_distr}{Logical indicating if returning the distributions of the permutation test.}

\item{seed}{A seed for replicability}

\item{verbose}{Print messages}

\item{sparse}{Using sparse or dense data from Seurat}
}
\value{
A data.table with the CCI interactions.
}
\description{
Run scDiffCom analysis
}
