% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdiffcom.R
\name{run_ORA}
\alias{run_ORA}
\title{Run over-representation analysis on the results from scDiffCom}
\usage{
run_ORA(
  scdiffcom_result,
  verbose = TRUE,
  logfc_threshold = scdiffcom_result$parameters$cutoff_logfc,
  categories = c("L_CELLTYPE", "R_CELLTYPE", "LR_CELLTYPE", "LR_NAME"),
  ora_types = c("UP", "DOWN", "DIFF", "FLAT")
)
}
\arguments{
\item{scdiffcom_result}{The list of results returned by run_scdiffcom}

\item{verbose}{Print messages}

\item{logfc_threshold}{Log fold-change threshold (in absolute value) to be used to define the categories of interest.
Default to cutoff_logfc used for the filtering analysis.}

\item{categories}{The categories over which the test is done}

\item{ora_types}{The type of regulation. Can be one or several of c("UP", "DOWN", "DIFF", "FLAT")}
}
\value{
A data.table
}
\description{
Run over-representation analysis on the results from scDiffCom
}
