% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_simple_analysis}
\alias{run_simple_analysis}
\title{Title}
\usage{
run_simple_analysis(
  expr_tr,
  metadata,
  cell_types,
  cond,
  LR_df,
  threshold,
  compute_fast
)
}
\arguments{
\item{expr_tr}{x}

\item{metadata}{x}

\item{cell_types}{x}

\item{cond}{x}

\item{LR_df}{x}

\item{threshold}{x}

\item{compute_fast}{x}
}
\value{
x
}
\description{
Title
}
