% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_stat_analysis}
\alias{run_stat_analysis}
\title{Title}
\usage{
run_stat_analysis(
  cci_dt_simple,
  expr_tr,
  metadata,
  cond_info,
  pp_LR,
  iterations,
  one_sided,
  return_distr = FALSE
)
}
\arguments{
\item{cci_dt_simple}{x}

\item{expr_tr}{x}

\item{metadata}{x}

\item{cond_info}{x}

\item{pp_LR}{x}

\item{iterations}{x}

\item{one_sided}{logical indicating if computing differential p-values from a one-sided or two-sided test.}

\item{return_distr}{x}
}
\value{
x
}
\description{
Title
}
