% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcom_analysis.R
\name{run_stat_iteration}
\alias{run_stat_iteration}
\title{Title}
\usage{
run_stat_iteration(expr_tr, metadata, template_cci_dt, pp_LR, cond_info)
}
\arguments{
\item{expr_tr}{x}

\item{metadata}{x}

\item{template_cci_dt}{x}

\item{pp_LR}{x}

\item{cond_info}{x}
}
\value{
x
}
\description{
Title
}
