% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RunORA}
\alias{RunORA}
\alias{RunORA,scDiffCom-method}
\title{Run over-representation analysis}
\usage{
RunORA(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS", "KEGG_PWS"),
  overwrite = TRUE,
  verbose = TRUE
)

\S4method{RunORA}{scDiffCom}(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS", "KEGG_PWS"),
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{scDiffCom} object}

\item{categories}{Name of the categories on which to perform ORA. Default
is \code{c("ER_CELLTYPES", "LRI", "GO_TERMS")}.}

\item{overwrite}{If \code{TRUE} (default), previous results are overwritten
in case they correspond to a category passed in \code{categories}.}

\item{verbose}{If \code{TRUE} (default) progress messages are printed.}
}
\value{
A scDiffCom object with updated \code{ora_table}.
}
\description{
Perform over-representation analysis (ORA) on a scDiffCom object. Note that
ORA is run automatically on default categories when calling
\code{run_interaction_analysis} or \code{FilterCCI}. This function may be
useful to perform ORA on other custom categories.
}
