% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RunORA}
\alias{RunORA}
\alias{RunORA.scDiffCom}
\alias{RunORA.scDiffComCombined}
\title{Perform over-representation analysis on detected intercellular communication patterns}
\usage{
RunORA(object, ...)

\method{RunORA}{scDiffCom}(
  object,
  categories = c("ER_CELLTYPES", "LR_GENES", "GO_TERMS"),
  overwrite = TRUE,
  stringent_or_default = "default",
  stringent_logfc_threshold = NULL,
  verbose = TRUE,
  ...
)

\method{RunORA}{scDiffComCombined}(
  object,
  categories = c("ER_CELLTYPES", "LR_GENES", "GO_TERMS"),
  overwrite = TRUE,
  stringent_or_default = "default",
  stringent_logfc_threshold = NULL,
  global = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An scDiffCom object previously returned by \code{run_interaction_analysis}.}

\item{...}{Arguments passed to other methods}

\item{categories}{A character vector specifying the categories on which to perform the analysis. One data.table is returned
for each category. Set to \code{c("ER_CELLTYPES", "LR_GENES", "GO_TERMS")} by default.}

\item{overwrite}{Should existing categories be overwriten in case they match with new categories?}

\item{stringent_or_default}{Should the default or more stringent ORA be performed?}

\item{stringent_logfc_threshold}{A more stringent logfc threshold compared to the one stored in \code{parameters}.
Set to \code{NULL} by default, namely no stringent ORA is done.}

\item{verbose}{Should messages be printed?}
}
\value{
An S4 object of class \code{scDiffCom} with updated slots \code{ora_defaut} (and potentially \code{ora_stringent}).
}
\description{
Over-representation analysis performed when an scDiffCom object contains results
 obtained from differential expression analysis.
}
