% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RunORA}
\alias{RunORA}
\alias{RunORA,scDiffCom-method}
\alias{RunORA,scDiffComCombined-method}
\title{xxx}
\usage{
RunORA(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS", "KEGG_PWS"),
  overwrite = TRUE,
  verbose = TRUE
)

\S4method{RunORA}{scDiffCom}(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS", "KEGG_PWS"),
  overwrite = TRUE,
  verbose = TRUE
)

\S4method{RunORA}{scDiffComCombined}(
  object,
  categories = c("ER_CELLTYPES", "LRI", "GO_TERMS", "KEGG_PWS"),
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{xxx}

\item{categories}{A character vector specifying the categories on which to perform the analysis. One data.table is returned
for each category. Set to \code{c("ER_CELLTYPES", "LRI", "GO_TERMS")} by default.}

\item{overwrite}{Should existing categories be overwritten in case they match with new categories?}

\item{verbose}{Should messages be printed?}
}
\value{
xxx
}
\description{
xxx
}
