% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{scDiffCom-class}
\alias{scDiffCom-class}
\title{The scDiffCom Class}
\description{
An object of this class stores the intercellular communication results
obtained by calling \code{run_interaction_analysis}.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used to build the object.}

\item{\code{cci_table_raw}}{Data.table with all potential CCIs induced from
the original Seurat object and the internal LRI database.}

\item{\code{cci_table_detected}}{Data.table with only the detected CCIs.}

\item{\code{ora_table}}{List of data.tables with the results of the
over-representation analysis.}

\item{\code{distributions}}{List of matrices with the null distributions for each
CCI.}
}}

