% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{scDiffCom-class}
\alias{scDiffCom-class}
\alias{scDiffCom}
\title{The scDiffCom Class}
\usage{
scDiffCom(
  parameters,
  cci_table_raw,
  cci_table_filtered,
  distributions,
  ora_tables
)
}
\description{
The scDiffCom Class is the class that stores the results of the
cell-cell interaction (CCI) analysis. Each object of this class is intended
to store the results of a single "tissue".
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of the parameters used to perform the CCI analysis.}

\item{\code{cci_table_raw}}{Data.table of all possible CCIs inferred from the single-cell data and Ligand-Receptor database.}

\item{\code{cci_table_filtered}}{Data.table of all detected CCIs obtained from the raw data.table after filtering.}

\item{\code{distributions}}{A list of matrices with the distributions returned by the permutation test(s).}

\item{\code{ora_tables}}{A list of data.tables with the results of the over-representation analysis for each specified category.}
}}

