% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{scDiffComBase-class}
\alias{scDiffComBase-class}
\title{The scDiffComBase Class}
\description{
The scDiffComBase class is a virtual class that provides a template
for the \code{scDiffCom} and \code{scDiffComCombined} classes.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{The list of parameters passed to
\code{run_interaction_analysis}.}

\item{\code{cci_table_raw}}{A data.table with all possible cell-cell
interactions (CCIs); namely with \eqn{n^2 \times m} rows,
where \eqn{n} is the number of cell-types and \eqn{m} the number
 of ligand-receptor pairs.}

\item{\code{cci_table_detected}}{A data.table with the detected CCIs obtained
from the raw data.table after filtering.}

\item{\code{ora_table}}{A data.table storing the results of the
over-representation analysis performed on the
 \code{cci_table_detected} table.}
}}

