% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{scDiffComBase-class}
\alias{scDiffComBase-class}
\alias{parameters,scDiffComBase-method}
\alias{get_cci_raw,scDiffComBase-method}
\alias{get_cci_detected,scDiffComBase-method}
\alias{get_ora_default,scDiffComBase-method}
\alias{get_ora_stringent,scDiffComBase-method}
\title{The scDiffComBase Class}
\usage{
\S4method{parameters}{scDiffComBase}(object)

\S4method{get_cci_raw}{scDiffComBase}(object)

\S4method{get_cci_detected}{scDiffComBase}(object)

\S4method{get_ora_default}{scDiffComBase}(object)

\S4method{get_ora_stringent}{scDiffComBase}(object)
}
\description{
The scDiffComBase class is a virtual class that provides a template
for the scDiffCom and scDiffComCombined classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{parameters}: Return scDiffCom \code{parameters}.

\item \code{get_cci_raw}: Return scDiffCom \code{cci_raw}.

\item \code{get_cci_detected}: Return scDiffCom \code{cci_detected}.

\item \code{get_ora_default}: Return scDiffCom \code{ora_default}.

\item \code{get_ora_stringent}: Return scDiffCom \code{ora_stringent}.
}}

\section{Slots}{

\describe{
\item{\code{parameters}}{A list of the parameters passed to \code{run_interaction_analysis}.}

\item{\code{cci_raw}}{A data.table of all possible cell-cell interactions (CCIs); namely with \eqn{n^2 \cdot m} rows,
where n is the number of cell-types and m the number of ligand-receptor pairs.}

\item{\code{cci_detected}}{A data.table of the detected CCIs obtained from the raw data.table after filtering.}

\item{\code{ora_default}}{A data.table storing the results of the over-representation analysis
performed on the \code{cci_detected} table.}

\item{\code{ora_stringent}}{Same as \code{ora_default} but with a more stringent logfc threshold to specifically focus on strong signals.}
}}

