% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{scDiffComCombined-class}
\alias{scDiffComCombined-class}
\title{The scDiffComCombined Class}
\description{
A scDiffComCombined object ...
}
\section{Slots}{

\describe{
\item{\code{parameters}}{A list of the parameters passed to \code{run_interaction_analysis}.}

\item{\code{cci_raw}}{A data.table of all possible cell-cell interactions (CCIs); namely with \eqn{n^2 \cdot m} rows,
where n is the number of cell-types and m the number of ligand-receptor pairs.}

\item{\code{cci_detected}}{A data.table of the detected CCIs obtained from the raw data.table after filtering.}

\item{\code{ora_default}}{A data.table storing the results of the over-representation analysis
performed on the \code{cci_detected} table.}

\item{\code{ora_stringent}}{Same as \code{ora_default} but with a more stringent logfc threshold to specifically focus on strong signals.}

\item{\code{combined_object}}{xxx}
}}

