% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_sample_tms_liver}
\alias{seurat_sample_tms_liver}
\title{A down-sampled Seurat object to use for testing and benchmarking}
\format{
An object of class Seurat.
}
\usage{
data(seurat_sample_tms_liver)
}
\description{
This Seurat object has been down-sampled from the original
Tabula Muris Senis Liver object. Pre-processing and normalization has
been performed before down-sampling. It contains 726 features (genes) and
468 samples (cells). It is only intended to be used for testing and
benchmarking and does not contain meaningful biological information.
}
\references{
\emph{A single-cell transcriptomic atlas characterizes
 ageing tissues in the mouse}, Tabula Muris Consortium (2020)
(\href{https://pubmed.ncbi.nlm.nih.gov/32669714/}{PMID: 32669714})
}
\keyword{datasets}
